requirements.txt
Flask==2.3.4
mysql-connector-python==9.0.0
﻿blinker==1.9.0
click==8.2.1
colorama==0.4.6
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
Werkzeug==3.1.3
